+++ comments.php
@@ -28,11 +28,38 @@ if ( ! comments_open() && get_comments_number() < 1 ) {
 
 if ( 'full-screen' === oceanwp_post_layout() ) {
 	$classes .= ' container';
-} ?>
+}
+
+// Comment form text attributes.
+$comment_logout_text = __( 'Log out of this account', 'oceanwp' );
+$comment_placeholder = __( 'Your Comment Here...', 'oceanwp' );
+$comment_profile_edit = __( 'Click to edit your profile', 'oceanwp' );
+
+// Get comment form position.
+$comment_position = apply_filters( 'ocean_comment_form_position', get_theme_mod( 'ocean_comment_form_position' ) );
+$comment_position = $comment_position ? $comment_position : 'after';
+
+// Comment form args.
+$args = array(
+	'must_log_in'           => '<p class="must-log-in">'.  sprintf( esc_html__( 'You must be %1$slogged in%2$s to post a comment.', 'oceanwp' ), '<a href="'. wp_login_url( apply_filters( 'the_permalink', get_permalink() ) ) .'">', '</a>' ) .'</p>',
+	'logged_in_as'          => '<p class="logged-in-as">'. esc_html__( 'Logged in as', 'oceanwp' ) .' <a href="'. admin_url( 'profile.php' ) .'">'. $user_identity .'</a>.<span class="screen-reader-text">'. $comment_profile_edit .'</span> <a href="' . wp_logout_url( get_permalink() ) .'" aria-label="'. esc_attr( $comment_logout_text ) .'">'. esc_html__( 'Log out', 'oceanwp' ) .'<span aria-hidden="true"> &raquo;</span></a></p>',
+	'comment_notes_before'  => false,
+	'comment_notes_after'   => false,
+	'comment_field'         => '<div class="comment-textarea"><label for="comment" class="screen-reader-text">'. esc_html__( 'Comment', 'oceanwp' ) . '</label><textarea name="comment" id="comment" cols="39" rows="4" tabindex="0" class="textarea-comment" placeholder="'. esc_attr( $comment_placeholder ) .'"></textarea></div>',
+	'id_submit'             => 'comment-submit',
+	'label_submit'          => esc_html__( 'Post Comment', 'oceanwp' ),
+);
+
+?>
 
 <section id="comments" class="<?php echo esc_attr( $classes ); ?>">
 
-	<?php // You can start editing here -- including this comment! ?>
+	<?php // You can start editing here -- including this comment!
+	// Display comment form if position set to before the comment list.
+	if ( $comment_position === 'before' ) {
+		comment_form( $args );
+	}
+	?>
 
 	<?php
 
@@ -105,24 +132,10 @@ if ( 'full-screen' === oceanwp_post_layout() ) {
 	<?php endif; // have_comments() ?>
 
 	<?php
-
-	// Comment form attributes.
-	$comment_logout_text = __( 'Log out of this account', 'oceanwp' );
-	$comment_placeholder = __( 'Your Comment Here...', 'oceanwp' );
-	$comment_profile_edit = __( '. Click to edit your profile', 'oceanwp' );
-
-	comment_form(
-		array(
-			/* translators: 1: Login URL 2: </a> */
-			'must_log_in'          => '<p class="must-log-in">' . sprintf( esc_html__( 'You must be %1$slogged in%2$s to post a comment.', 'oceanwp' ), '<a href="' . wp_login_url( apply_filters( 'the_permalink', get_permalink() ) ) . '">', '</a>' ) . '</p>',
-			'logged_in_as'         => '<p class="logged-in-as">' . esc_html__( 'Logged in as', 'oceanwp' ) . ' <a href="' . admin_url( 'profile.php' ) . '" aria-label="'. esc_attr( $user_identity . $comment_profile_edit ) .'">' . $user_identity . '</a>. <a href="' . wp_logout_url( get_permalink() ) . '" aria-label="' . esc_attr( $comment_logout_text ) . '">' . esc_html__( 'Log out', 'oceanwp' ) . '<span aria-hidden="true"> &raquo;</span>' .'</a></p>',
-			'comment_notes_before' => false,
-			'comment_notes_after'  => false,
-			'comment_field'        => '<div class="comment-textarea"><label for="comment" class="screen-reader-text">' . esc_html__( 'Enter your text to leave a comment', 'oceanwp' ) . '</label><textarea name="comment" id="comment" cols="39" rows="4" tabindex="0" class="textarea-comment" placeholder="' . esc_attr( $comment_placeholder ) . '"></textarea></div>',
-			'id_submit'            => 'comment-submit',
-			'label_submit'         => esc_html__( 'Post Comment', 'oceanwp' ),
-		)
-	);
+	// Display comment form if position set to after the comment list (default setting).
+	if ( $comment_position === 'after' ) {
+		comment_form( $args );
+	}
 	?>
 
 </section><!-- #comments -->